+++ translated_fields/fields.py
@@ -1,4 +1,6 @@
+import inspect
 import re
+import warnings
 
 from django.conf import settings
 from django.db.models import Field
@@ -36,6 +38,20 @@ def translated_attributes(*names, attrgetter=translated_attrgetter):
     return decorator
 
 
+def _optional_keywords(fn, *args, **kwargs):
+    params = inspect.signature(fn).parameters
+    if kwargs.keys() - params.keys():
+        warnings.warn(
+            "%s has unsupported arguments: %s"
+            % (
+                getattr(fn, "__name__", fn),
+                ", ".join(sorted(kwargs.keys() - params.keys())),
+            ),
+            DeprecationWarning,
+        )
+    return fn(*args, **{key: value for key, value in kwargs.items() if key in params})
+
+
 class TranslatedField(object):
     def __init__(
         self, field, specific=None, *, languages=None, attrgetter=None, attrsetter=None
@@ -67,8 +83,9 @@ class TranslatedField(object):
         setattr(cls, name, self)
         self.fields = fields
         self.short_description = kwargs.get("verbose_name", name)
-        self._getter = self._attrgetter(name)
-        self._setter = self._attrsetter(name)
+
+        self._getter = _optional_keywords(self._attrgetter, name, field=self)
+        self._setter = _optional_keywords(self._attrsetter, name, field=self)
 
     def __get__(self, obj, objtype=None):
         if obj is None: