+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.8`_ (2019-06-26)
+~~~~~~~~~~~~~~~~~~~
+
 - Added an optional ``field`` argument to the ``attrgetter`` and
   ``attrsetter`` functions.
 - Added a utils module intended to contain common applications of
@@ -112,4 +116,5 @@ Change log
 .. _0.5: https://github.com/matthiask/django-translated-fields/compare/0.4...0.5
 .. _0.6: https://github.com/matthiask/django-translated-fields/compare/0.5...0.6
 .. _0.7: https://github.com/matthiask/django-translated-fields/compare/0.6...0.7
-.. _Next version: https://github.com/matthiask/django-translated-fields/compare/0.7...master
+.. _0.8: https://github.com/matthiask/django-translated-fields/compare/0.7...0.8
+.. _Next version: https://github.com/matthiask/django-translated-fields/compare/0.8...master